refactor: extract PlayerSelectorStrip from Game Tracker
