+++ src/pages/GameTracker.tsx
@@ -7,24 +7,10 @@ import { buildPeriodSegmentLabels, getBonusFoulCountForPeriod } from '../lib/tea
 import type { BasketballTeamStatsConfig, Player, StatAction, StatCategory } from '../types'
 import Scoreboard from '../components/Scoreboard'
 import StatButton from '../components/StatButton'
+import PlayerSelectorStrip from '../components/PlayerSelectorStrip'
 import PeriodToggle from '../components/team-stats/PeriodToggle'
 import BasketballBonusIndicator from '../components/team-stats/BasketballBonusIndicator'
-import {
-  isTeamPseudoPlayer,
-  playersWithTeamPlaceholders,
-  TEAM_PLAYER_HOME_ID,
-  TEAM_PLAYER_OPP_ID,
-} from '../lib/teamPlayers'
-
-function sortTeamPlayersFirst(players: Player[]): Player[] {
-  const teams = players.filter(isTeamPseudoPlayer)
-  const home = teams.find(p => p.id === TEAM_PLAYER_HOME_ID || p.teamSide === 'home')
-  const opp = teams.find(p => p.id === TEAM_PLAYER_OPP_ID || p.teamSide === 'opponent')
-  const restTeam = teams.filter(p => p !== home && p !== opp)
-  const individuals = players.filter(p => !isTeamPseudoPlayer(p))
-  const orderedTeams = [home, opp, ...restTeam].filter(Boolean) as Player[]
-  return [...orderedTeams, ...individuals]
-}
+import { playersWithTeamPlaceholders } from '../lib/teamPlayers'
 
 /** Strip `_pN` suffix for config lookup (undo log stores scoped ids). */
 function baseStatId(statId: string): string {
@@ -149,9 +135,6 @@ export default function GameTracker() {
     dispatch({ type: 'SET_PLAYERS', players: nextPlayers })
   }, [sport, gameInfo, players, dispatch])
 
-  const selectorPlayers = useMemo(() => sortTeamPlayersFirst(players), [players])
-  const teamSelectorCount = selectorPlayers.filter(isTeamPseudoPlayer).length
-
   if (!sport || !gameInfo || players.length === 0) {
     navigate('/')
     return null
@@ -258,59 +241,17 @@ export default function GameTracker() {
         )}
       </div>
 
-      <div className="px-3 py-2 max-w-lg mx-auto w-full">
-        <div className="flex gap-2 overflow-x-auto pb-1 scrollbar-hide items-stretch">
-          {selectorPlayers.map((player, index) => {
-            const isTeam = isTeamPseudoPlayer(player)
-            const showDivider = isTeam && index === teamSelectorCount - 1 && teamSelectorCount > 0
-            const isActive = player.id === activePlayer.id
+      <PlayerSelectorStrip
+        players={players}
+        activePlayerId={activePlayer.id}
+        onSelectPlayer={playerId => dispatch({ type: 'SET_ACTIVE_PLAYER', playerId })}
+        activeBgClass={sport.theme.bg}
+        onAddPlayer={() => setShowAddPlayer(!showAddPlayer)}
+      />
 
-            return (
-              <div key={player.id} className="flex flex-shrink-0 items-stretch gap-2">
-                <button
-                  onClick={() => dispatch({ type: 'SET_ACTIVE_PLAYER', playerId: player.id })}
-                  title={player.name}
-                  className={`
-                    flex-shrink-0 px-3 py-2 rounded-xl text-sm font-semibold max-w-[10.5rem]
-                    transition-all duration-150 active:scale-95 text-left
-                    ${isTeam
-                      ? isActive
-                        ? `bg-gradient-to-br from-slate-600 to-slate-800 text-white shadow-md ring-2 ring-white/30`
-                        : `bg-gradient-to-br from-slate-100 to-slate-200/90 text-slate-800 border border-slate-300/80 shadow-sm`
-                      : isActive
-                        ? `${sport.theme.bg} text-white shadow-md`
-                        : 'bg-white text-slate-600 border border-slate-200'
-                    }
-                  `}
-                >
-                  <span className={isTeam ? 'opacity-90' : 'opacity-70'}>
-                    {isTeam ? '★' : `#${player.number || '?'}`}
-                  </span>{' '}
-                  <span className="line-clamp-2 break-words">
-                    {isTeam ? player.name : player.name.split(' ')[0]}
-                  </span>
-                </button>
-                {showDivider && (
-                  <div
-                    className="w-px self-stretch min-h-[2.5rem] bg-slate-300 shrink-0"
-                    aria-hidden
-                  />
-                )}
-              </div>
-            )
-          })}
-          <button
-            onClick={() => setShowAddPlayer(!showAddPlayer)}
-            className="flex-shrink-0 w-10 h-10 rounded-xl bg-white border-2 border-dashed
-                       border-slate-300 text-slate-400 text-xl font-bold
-                       active:scale-95 transition-transform flex items-center justify-center"
-          >
-            +
-          </button>
-        </div>
-
-        {showAddPlayer && (
-          <div className="card mt-2 flex gap-2 items-end">
+      {showAddPlayer && (
+        <div className="px-3 max-w-lg mx-auto w-full">
+          <div className="card mb-2 flex gap-2 items-end">
             <input
               type="text"
               value={newNumber}
@@ -336,8 +277,8 @@ export default function GameTracker() {
               Add
             </button>
           </div>
-        )}
-      </div>
+        </div>
+      )}
 
       {showPeriodToggle && teamRules && (
         <div className="px-3 max-w-lg mx-auto w-full">
